fix:  delete old breadcrumbs and replace with bucket
