+++ ui/app/components/breadcrumbs.js
@@ -2,9 +2,9 @@ import Component from '@glimmer/component';
 import { inject as service } from '@ember/service';
 
 export default class Breadcrumbs extends Component {
-  @service bucket;
+  @service breadcrumbs;
 
   get crumbs() {
-    return this.bucket.crumbs;
+    return this.breadcrumbs.crumbs;
   }
 }
